service-unit:refactor: move config to secondary adapter
