+++ service-unit/stateless/internal/application/core/initialization/configloader.go
@@ -2,7 +2,7 @@ package initialization
 
 import (
 	model "github.com/hanapedia/the-bench/the-bench-operator/api/v1"
-	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure/config"
+	"github.com/hanapedia/the-bench/service-unit/stateless/internal/infrastructure_new/adapters/secondary/config"
 )
 
 // GetConfig parses service-unit config via given formant
